pyproject.toml
[project]
name = "bounty_hunter"
version = "0.1.0"
description = "FastAPI with endpoints to scrap and answer natural language questions about local Bay Area businesses."
authors = [{ name = "dychen", email = "dychen93@hotmail.com" }]
readme = "README.md"
requires-python = ">=3.12, <3.13"

[dependency-groups]
main = [
    "fastapi[standard] >=0.115.5",
    "pydantic >=2.9.2",
    "pydantic-settings >=2.6.1",
    "httpx>=0.28.1",
    "pandas>=2.2.3",
    "sqlalchemy>=2.0.41",
    "aiosqlite>=0.21.0",
    "google-genai>=1.16.1",
    "tenacity>=9.1.2",
    "async-lru>=2.0.5",
    "aiofiles>=24.1.0",
]
dev = [
    "mypy>=1.13.0",
 "pandas-stubs>=2.2.3.250308",
 "ruff >=0.9.9",
]

[tool.uv]
default-groups = ["main", "dev"]

[tool.mypy]
plugins = "pydantic.mypy"
